Module aws_sdk_ebs::error

Modules

access_denied_exception

See AccessDeniedException

concurrent_limit_exceeded_exception

See ConcurrentLimitExceededException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

request_throttled_exception

See RequestThrottledException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

CompleteSnapshotError

ConcurrentLimitExceededException

You have reached the limit for concurrent API requests. For more information, see

ConflictException

The request uses the same client token as a previous, but non-identical

GetSnapshotBlockError

InternalServerException

An internal error has occurred.

ListChangedBlocksError

ListSnapshotBlocksError

PutSnapshotBlockError

RequestThrottledException

The number of API requests has exceed the maximum allowed API request throttling

ResourceNotFoundException

The specified resource does not exist.

ServiceQuotaExceededException

Your current service quotas do not allow you to perform this action.

StartSnapshotError

ValidationException

The input fails to satisfy the constraints of the EBS direct APIs.

Enums

CompleteSnapshotErrorKind

GetSnapshotBlockErrorKind

ListChangedBlocksErrorKind

ListSnapshotBlocksErrorKind

PutSnapshotBlockErrorKind

StartSnapshotErrorKind