Struct aws_sdk_ebs::input::PutSnapshotBlockInput

#[non_exhaustive]
pub struct PutSnapshotBlockInput {
    pub snapshot_id: Option<String>,
    pub block_index: Option<i32>,
    pub block_data: ByteStream,
    pub data_length: Option<i32>,
    pub progress: Option<i32>,
    pub checksum: Option<String>,
    pub checksum_algorithm: Option<ChecksumAlgorithm>,
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

snapshot_id: Option<String>

The ID of the snapshot.

block_index: Option<i32>

The block index of the block in which to write the data. A block index is a logical index in units of 512 KiB blocks. To identify the block index, divide the logical offset of the data in the logical volume by the block size (logical offset of data/524288). The logical offset of the data must be 512 KiB aligned.

block_data: ByteStream

The data to write to the block.

The block data is not signed as part of the Signature Version 4 signing process. As a result, you must generate and provide a Base64-encoded SHA256 checksum for the block data using the x-amz-Checksum header. Also, you must specify the checksum algorithm using the x-amz-Checksum-Algorithm header. The checksum that you provide is part of the Signature Version 4 signing process. It is validated against a checksum generated by Amazon EBS to ensure the validity and authenticity of the data. If the checksums do not correspond, the request fails. For more information, see Using checksums with the EBS direct APIs in the Amazon Elastic Compute Cloud User Guide.

data_length: Option<i32>

The size of the data to write to the block, in bytes. Currently, the only supported size is 524288.

Valid values: 524288

progress: Option<i32>

The progress of the write process, as a percentage.

checksum: Option<String>

A Base64-encoded SHA256 checksum of the data. Only SHA256 checksums are supported.

checksum_algorithm: Option<ChecksumAlgorithm>

The algorithm used to generate the checksum. Currently, the only supported algorithm is SHA256.

Implementations

impl PutSnapshotBlockInput

pub fn make_operation(
self,
_config: &Config
) -> Result<Operation<PutSnapshotBlock, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<PutSnapshotBlock>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture PutSnapshotBlockInput