Module aws_sdk_ebs::model

Modules

block

See Block

changed_block

See ChangedBlock

tag

See Tag

Structs

Block

A block of data in an Amazon Elastic Block Store snapshot.

ChangedBlock

A block of data in an Amazon Elastic Block Store snapshot that is different from another snapshot of

Tag

Describes a tag.

Enums

AccessDeniedExceptionReason

ChecksumAggregationMethod

ChecksumAlgorithm

RequestThrottledExceptionReason

ResourceNotFoundExceptionReason

ServiceQuotaExceededExceptionReason

Status

ValidationExceptionReason