Struct aws_sdk_ebs::ByteStream

pub struct ByteStream(_);

Stream of binary data

ByteStream wraps a stream of binary data for ease of use.

Getting data out of a ByteStream

ByteStream provides two primary mechanisms for accessing the data:

With .collect(): .collect() reads the complete ByteStream into memory and stores it in AggregatedBytes, a non-contiguous ByteBuffer.

use smithy_http::byte_stream::{ByteStream, AggregatedBytes};
use smithy_http::body::SdkBody;
use bytes::Buf;
async fn example() {
   let stream = ByteStream::new(SdkBody::from("hello! This is some data"));
   // Load data from the stream into memory:
   let data = stream.collect().await.expect("error reading data");
   // collect returns a `bytes::Buf`:
   println!("first chunk: {:?}", data.chunk());
}

Via impl Stream: Note: An import of StreamExt is required to use try_next(). For use-cases where holding the entire ByteStream in memory is unecessary, use the Stream implementation:

use smithy_http::byte_stream::{ByteStream, AggregatedBytes, Error};
use smithy_http::body::SdkBody;
use tokio_stream::StreamExt;

async fn example() -> Result<(), Error> {
   let mut stream = ByteStream::from(vec![1, 2, 3, 4, 5, 99]);
   let mut digest = crc32::Digest::new();
   while let Some(bytes) = stream.try_next().await? {
       digest.write(&bytes);
   }
   println!("digest: {}", digest.finish());
   Ok(())
}

Getting data into a ByteStream

ByteStreams can be created in one of three ways:

From in-memory binary data: ByteStreams created from in-memory data are always retryable. Data will be converted into Bytes enabling a cheap clone during retries.

use bytes::Bytes;
use smithy_http::byte_stream::ByteStream;
let stream = ByteStream::from(vec![1,2,3]);
let stream = ByteStream::from(Bytes::from_static(b"hello!"));

From a file: ByteStreams created from a path can be retried. A new file descriptor will be opened if a retry occurs.

use smithy_http::byte_stream::ByteStream;
let stream = ByteStream::from_path("big_file.csv");

From an SdkBody directly: For more advanced / custom use cases, a ByteStream can be created directly from an SdkBody. When created from an SdkBody, care must be taken to ensure retriability. An SdkBody is retryable when constructured from in-memory data or when using SdkBody::retryable.

use smithy_http::byte_stream::ByteStream;
use smithy_http::body::SdkBody;
use bytes::Bytes;
let (mut tx, channel_body) = hyper::Body::channel();
// this will not be retryable because the SDK has no way to replay this stream
let stream = ByteStream::new(SdkBody::from(channel_body));
tx.send_data(Bytes::from_static(b"hello world!"));
tx.send_data(Bytes::from_static(b"hello again!"));
// NOTE! You must ensure that `tx` is dropped to ensure that EOF is sent

Implementations

impl ByteStream

pub fn new(body: SdkBody) -> ByteStream

pub fn from_static(bytes: &'static [u8]) -> ByteStream

pub fn into_inner(self) -> SdkBody

Consumes the ByteStream, returning the wrapped SdkBody

pub async fn collect(self) -> Result<AggregatedBytes, Error>

Read all the data from this ByteStream into memory

If an error in the underlying stream is encountered, ByteStreamError is returned.

Data is read into an AggregatedBytes that stores data non-contiguously as it was received over the network. If a contiguous slice is required, use into_bytes().

use bytes::Bytes;
use smithy_http::body;
use smithy_http::body::SdkBody;
use smithy_http::byte_stream::{ByteStream, Error};
async fn get_data() {
    let stream = ByteStream::new(SdkBody::from("hello!"));
    let data: Result<Bytes, Error> = stream.collect().await.map(|data| data.into_bytes());
}

pub async fn from_path(path: impl AsRef<Path>) -> Result<ByteStream, Error>

Create a ByteStream that streams data from the filesystem

This function creates a retryable ByteStream for a given path. The returned ByteStream will provide a size hint when used as an HTTP body. If the request fails, the read will begin again by reloading the file handle.

Warning

The contents of the file MUST not change during retries. The length & checksum of the file will be cached. If the contents of the file change, the operation will almost certainly fail.

Furthermore, a partial write MAY seek in the file and resume from the previous location.

Example

use smithy_http::byte_stream::ByteStream;
use std::path::Path;
 async fn make_bytestream() -> ByteStream {
    ByteStream::from_path("docs/rows.csv").await.expect("file should be readable")
}

pub async fn from_file(file: File) -> Result<ByteStream, Error>

Create a ByteStream from a file

NOTE: This will NOT result in a retryable ByteStream. For a ByteStream that can be retried in the case of upstream failures, use ByteStream::from_path